Tested with Python 3.12.0b1, assume it will work with the Python 3.12 release as well.
Note: Python 3.12.0b1 breaks the current release version of lxml.
This is expecteded to be fixed in lxml 5.0.0 but that version has not
been released yet, so I could not run any of the tests that depend on
lxml with Python 3.12.0b1.  But I am confident that there is no issue
related to python-icat with that, so I tick 3.12 to be working
nevertheless.